fix: moved getLogger to own file
