+++ control/startup.py
@@ -7,9 +7,10 @@ May be ran as script
 """
 import os
 
-import filehandler as fh
+from myLogger import getLogger
+from tweakable import BALANCE_CATEGORIES, RESULT_CATEGORIES
 
-from ..main import getLogger
+from . import filehandler as fh
 
 
 def isFirstTime():
@@ -34,6 +35,30 @@ def makeFiles():
 
     log.info("files created successfully...")
 
+
+    fh.writeFile("balances.json", resetBalanceData())
+    log.info("added balance values..")
+
+    fh.writeFile("results.json", resetResultData())
+    log.info("added resultvalues..")
+
+
+
+def resetResultData():
+    """Converts the list to a dict where list
+    becomes keys with value 0"""
+    resData = {}
+    for resCat in RESULT_CATEGORIES():
+        resData[resCat] = 0
+    return resData
+
+
+def resetBalanceData():
+    balData = {}
+    balCats = BALANCE_CATEGORIES()
+    for balCat in balCats: 
+        balData[balCat] = {"init_value": balCats[balCat], "value": 0 }
+    return balData
 
 if __name__ == "__main__":
    makeFiles()
