+++ notes.md
@@ -16,7 +16,7 @@ tweakable -> startup/filecreation -> windowstartup -> input -> writeFile
 ### branch: feat: data setup
 - feat: make files startup(DONE)
 - feat: setup data files (startup)
-  1. finish tweakable.py
+  1. finish tweakable.py (DONE)
   2. add to makeFiles()
 - feat: make dictmapper
